requirements.txt
# Backend & Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Web Scraping
selenium==4.15.0
beautifulsoup4==4.12.0
requests==2.31.0

# Data Processing
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.3.2

# Deep Learning
tensorflow==2.13.0
torch==2.1.0
torchvision==0.16.0

# NLP & Text Processing
nltk==3.8.1
spacy==3.7.0
textblob==0.17.0

# Machine Learning Utilities
joblib==1.3.2
pickle-mixin==1.0.0

# Email & Domain Validation
email-validator==2.1.0
validators==0.22.0
python-whois==0.9.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.0
